requirements.txt
google-generativeai
requests
requests[socks]
PySocks
rich
deepseek